+++ lumin/nn/models/blocks/head.py
@@ -9,64 +9,56 @@ import torch.nn as nn
 from torch.tensor import Tensor
 import torch
 
-from ..helpers import Embedder
+from ..helpers import CatEmbedder
 from ..initialisations import lookup_normal_init
 from ....plotting.plot_settings import PlotSettings
 from ....plotting.interpretation import plot_embedding
+from .abs_block import AbsBlock
 
 
-class CatEmbHead(nn.Module):
+class AbsHead(AbsBlock):
+    def __init__(self, n_cont_in:int, cat_embedder:Optional[CatEmbedder]=None, 
+                 lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init, freeze:bool=False):
+        super().__init__(lookup_init=lookup_init, freeze=freeze)
+        self.n_cont_in,self.cat_embedder = n_cont_in,cat_embedder
+
+
+class CatEmbHead(AbsHead):
     r'''
     Standard model head for columnar data.
     Provides inputs for continuous features and embedding matrices for categorical inputs, and uses a dense layer to upscale to width of network body.
     Designed to be passed as a 'head' to :class:ModelBuilder.
     Supports batch normalisation and dropout (at separate rates for pre-dense continuous, pre-dense categorical embeddings, and post dense).
     Continuous features are expected to be the first n_cont_in columns of input tensors and categorical features the remaining columns.
-    Embedding arguments for categorical features are set using a :class:Embedder.
+    Embedding arguments for categorical features are set using a :class:CatEmbedder.
 
     Arguments:
         n_cont_in: number of continuous inputs to expect
         do_cont: if not None will add a dropout layer with dropout rate do acting on the continuous inputs prior to concatination wih the categorical embeddings
         do_cat: if not None will add a dropout layer with dropout rate do acting on the categorical embeddings prior to concatination wih the continuous inputs
-        cat_embedder: :class:Embedder providing details of how to embed categorical inputs
+        cat_embedder: :class:CatEmbedder providing details of how to embed categorical inputs
         lookup_init: function taking choice of activation function, number of inputs, and number of outputs an returning a function to initialise layer weights.
         freeze: whether to start with module parameters set to untrainable
 
     Examples::
         >>> head = CatEmbHead(n_cont_in=30)
-        >>> head = CatEmbHead(n_cont_in=25, cat_embedder=Embedder.from_fy(train_fy))
-        >>> head = CatEmbHead(n_cont_in=25, cat_embedder=Embedder.from_fy(train_fy), do_cont=0.1, do_cat=0.05)
-        >>> head = CatEmbHead(n_cont_in=25, cat_embedder=Embedder.from_fy(train_fy), lookup_init=lookup_uniform_init)
+        >>> head = CatEmbHead(n_cont_in=25, cat_embedder=CatEmbedder.from_fy(train_fy))
+        >>> head = CatEmbHead(n_cont_in=25, cat_embedder=CatEmbedder.from_fy(train_fy), do_cont=0.1, do_cat=0.05)
+        >>> head = CatEmbHead(n_cont_in=25, cat_embedder=CatEmbedder.from_fy(train_fy), lookup_init=lookup_uniform_init)
     '''
 
-    def __init__(self, n_cont_in:int, do_cont:float=0, do_cat:float=0, cat_embedder:Optional[Embedder]=None, 
+    def __init__(self, n_cont_in:int, do_cont:float=0, do_cat:float=0, cat_embedder:Optional[CatEmbedder]=None, 
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init, freeze:bool=False):
-        super().__init__()
-        self.n_cont_in,self.do_cont,self.do_cat,self.cat_embedder,self.lookup_init,self.freeze = n_cont_in,do_cont,do_cat,cat_embedder,lookup_init,freeze
-        if self.cat_embedder is None: self.cat_embedder = Embedder([], [])
-        if self.cat_embedder.n_cat_in > 0: 
+        super().__init__(n_cont_in=n_cont_in, cat_embedder=cat_embedder, lookup_init=lookup_init, freeze=freeze)
+        self.do_cont,self.do_cat, = do_cont,do_cat
+        if self.cat_embedder is None: self.cat_embedder = CatEmbedder([], [])
+        if self.cat_embedder.n_cat_in > 0:
             self.embeds = nn.ModuleList([nn.Embedding(ni, no) for _, ni, no in self.cat_embedder])
             if self.cat_embedder.emb_load_path is not None: self._load_embeds()
             if self.do_cat   > 0: self.emb_do     = nn.Dropout(self.do_cat)
         self.n_out = self.n_cont_in if self.cat_embedder.n_cat_in == 0 else self.n_cont_in+np.sum(self.cat_embedder.emb_szs)
         if self.do_cont  > 0: self.cont_in_do = nn.Dropout(self.do_cont)
         if self.freeze: self.freeze_layers()
-    
-    def __getitem__(self, key:int) -> nn.Module: return self.layers[key]
-
-    def freeze_layers(self) -> None:
-        r'''
-        Make parameters untrainable
-        '''
-
-        for p in self.parameters(): p.requires_grad = False
-    
-    def unfreeze_layers(self) -> None:
-        r'''
-        Make parameters trainable
-        '''
-
-        for p in self.parameters(): p.requires_grad = True
         
     def forward(self, x_in:Tensor) -> Tensor:
         if self.cat_embedder.n_cat_in > 0: