+++ lumin/nn/models/helpers.py
@@ -1,14 +1,26 @@
 from typing import List, Tuple, Optional, Union
 from pathlib import Path
 import numpy as np
+import warnings
 
 from ..data.fold_yielder import FoldYielder
 
 
-class Embedder():
+def Embedder(cat_names:List[str], cat_szs:List[int], emb_szs:Optional[List[int]]=None, max_emb_sz:int=50,
+             emb_load_path:Optional[Union[Path,str]]=None):
+    r'''
+    Depreciated: renamed to plot_rank_order_dendrogram
+    '''
+    
+    # XXX Remove in v0.3
+    warnings.warn('''Embedder has been renamed to CatEmbedder. Embedder is now depreciated and will be removed in v0.3''')
+    return CatEmbedder(cat_names=cat_names, cat_szs=cat_szs, emb_szs=emb_szs, max_emb_sz=max_emb_sz, emb_load_path=emb_load_path)
+
+
+class CatEmbedder():
     r'''
     Helper class for embedding categorical features. Designed to be passed to :class:ModelBuilder.
-    Note that the classmethod :meth:from_fy may be used to instantiate an :class:Embedder from a :class:FoldYielder.
+    Note that the classmethod :meth:from_fy may be used to instantiate an :class:CatEmbedder from a :class:FoldYielder.
 
     Arguments:
         cat_names: list of names of catgorical features in order in which they will be passed as inputs columns
@@ -18,9 +30,9 @@ class Embedder():
         emb_load_path: if not None, will cause :class:ModelBuilder to attempt to load pretrained embeddings from path
 
     Examples::
-        >>> cat_embedder = Embedder(cat_name=['n_jets', 'channel'], cat_szs=[5, 3])
-        >>> cat_embedder = Embedder(cat_name=['n_jets', 'channel'], cat_szs=[5, 3], emb_szs=[2, 2])
-        >>> cat_embedder = Embedder(cat_name=['n_jets', 'channel'], cat_szs=[5, 3], emb_szs=[2, 2], emb_load_path=Path('weights'))
+        >>> cat_embedder = CatEmbedder(cat_names=['n_jets', 'channel'], cat_szs=[5, 3])
+        >>> cat_embedder = CatEmbedder(cat_names=['n_jets', 'channel'], cat_szs=[5, 3], emb_szs=[2, 2])
+        >>> cat_embedder = CatEmbedder(cat_names=['n_jets', 'channel'], cat_szs=[5, 3], emb_szs=[2, 2], emb_load_path=Path('weights'))
     '''
 
     # TODO: load pretrained embeddings to check sizes
@@ -50,7 +62,7 @@ class Embedder():
     @classmethod
     def from_fy(cls, fy:FoldYielder,  emb_szs:Optional[List[int]]=None, max_emb_sz:int=50, emb_load_path:Optional[Union[Path,str]]=None):
         r'''
-        Instantiate an :class:Embedder from a :class:FoldYielder, i.e. avoid having to pass cat_names and cat_szs.
+        Instantiate an :class:CatEmbedder from a :class:FoldYielder, i.e. avoid having to pass cat_names and cat_szs.
         
         Arguments:
             fy: :class:FoldYielder with training data
@@ -59,12 +71,12 @@ class Embedder():
             emb_load_path: if not None, will cause :class:ModelBuilder to attempt to load pretrained embeddings from path
 
         Returns:
-            :class:Embedder
+            :class:CatEmbedder
 
         Examples::
-            >>> cat_embedder = Embedder.from_fy(train_fy)
-            >>> cat_embedder = Embedder.from_fy(train_fy, emb_szs=[2, 2])
-            >>> cat_embedder = Embedder.from_fy(train_fy, emb_szs=[2, 2], emb_load_path=Path('weights'))
+            >>> cat_embedder = CatEmbedder.from_fy(train_fy)
+            >>> cat_embedder = CatEmbedder.from_fy(train_fy, emb_szs=[2, 2])
+            >>> cat_embedder = CatEmbedder.from_fy(train_fy, emb_szs=[2, 2], emb_load_path=Path('weights'))
         '''
 
         cat_names = fy.cat_feats